Resolve #11: Fixing fox's Komi values
This is just the functional part and two test files. Test cases coming
next as well as documentation.

Adjusting Komi on fox is added as an option you can pass to ParseToken.

This handles fox's komi in 3 cases by following kaTrain's fixup:
1. Handicap games it adjusts to 0.5 Komi
2. Even games using chinese rules it adjusts to 7.5 Komi (from 375)
3. From other rulesets it adjusts to (6.5 from 325)

Note: probably a bug here as some games have 0 as Komi and we shouldn't
add Komi in those cases.