Revise adjustments to Fox Komi
After finding a number of test cases, I think this is more likely to get
the various scenarios across handicap stones, rulesets, and what not
correct.

Update test cases to cover.

Added documentation.

Handling komi for fox is trickier than it seems. I settled on detecting
actual handicap stones to determine 0.5 scenarios and then used a
combination of rulesets and handicap setting to determine between 6.5
and 7.5 handicap.

Note: The "Better" way for some of this would be to walk the nodes
searching for the data we see. IE instead of nesting into the data
structure directly in search of Fox's AB nodes for adding black stones,
we should walk the tree searching for that node. However, since this
solution is about fixing up a specific go client I went with what is
faster.

For the record some servers like Pandanet (IGS) do put their handicap
stones with the rest of the game metadata in the root node and not in
the first child.

The logic is to assume a common handicap, 6.5 which is used on many game
formats and servers.

Then I check for fox specific scenarios around handicap and different
rulesets. After that, I use the presence of handicap stones to figure
out if the game should be 0.5 handicap and I assume the actual presence
of handicap stones trumps everything else on Fox.